Rename class from ClimateChamberWrapper to ClimateChamberControl
